refactor(axvirtio-blk): move parse_virtio_block_header to device.rs
- Move parse_virtio_block_header function from queue/mod.rs to device.rs
- Update related imports and usages
- This change prepares for implementing header parsing in the transport layer
