+++ axvirtio-common/src/queue/mod.rs
@@ -4,7 +4,7 @@ mod used;
 
 pub use available::{AvailableRing, VirtQueueAvail};
 pub use descriptor::{DescriptorTable, VirtQueueDesc};
-use log::{trace, warn};
+use log::trace;
 pub use used::{UsedRing, VirtQueueUsed, VirtqUsedElem};
 
 use crate::{
@@ -48,7 +48,7 @@ pub struct VirtioQueue<T: AddressTranslator + Clone> {
     /// Queue size
     pub size: u16,
     /// Descriptor table
-    desc_table: Option<DescriptorTable<T>>,
+    pub desc_table: Option<DescriptorTable<T>>,
     /// Available ring
     avail_ring: Option<AvailableRing<T>>,
     /// Used ring
@@ -259,51 +259,6 @@ impl<T: AddressTranslator + Clone> VirtioQueue<T> {
         if let Some(ref desc_table) = self.desc_table {
             let descriptors = desc_table.follow_chain(head_index)?;
             Ok(descriptors.len() >= min_length)
-        } else {
-            Err(VirtioError::QueueNotReady)
-        }
-    }
-
-    /// Parse VirtIO block header
-    pub fn parse_virtio_block_header(&self, head_index: u16) -> VirtioResult<VirtioBlockHeader> {
-        if let Some(ref desc_table) = self.desc_table {
-            let descriptors = desc_table.follow_chain(head_index)?;
-            if descriptors.is_empty() {
-                return Err(VirtioError::InvalidDescriptor);
-            }
-
-            // Get the first descriptor which should contain the request header
-            let header_desc = &descriptors[0];
-
-            // Validate that the first descriptor is readable (not write-only)
-            if header_desc.is_write() {
-                warn!("Request header descriptor should not be write-only");
-                return Err(VirtioError::InvalidDescriptor);
-            }
-
-            // Check if the descriptor is large enough to contain the header
-            if header_desc.len < VirtioBlockHeader::SIZE {
-                warn!(
-                    "Request header descriptor too small: {} bytes, need {} bytes",
-                    header_desc.len,
-                    VirtioBlockHeader::SIZE
-                );
-                return Err(VirtioError::InvalidDescriptor);
-            }
-
-            // Read the header from guest memory
-            let header_addr = header_desc.guest_addr();
-
-            // Use the structured header reading
-            let header = VirtioBlockHeader::read_from_guest(header_addr, self.accessor.clone())?;
-
-            trace!(
-                "Parsed VirtIO block header: type={}, sector={}",
-                header.request_type,
-                header.sector
-            );
-
-            Ok(header)
         } else {
             Err(VirtioError::QueueNotReady)
         }
